Moved summary to separate component
